+++ nextcloudappstore/core/github.py
@@ -1,8 +1,10 @@
 from itertools import chain, takewhile
-from typing import Iterable
+from typing import Iterable, List
 
 import requests
 from semantic_version import Version
+
+from nextcloudappstore.core.models import NextcloudRelease
 
 
 class GitHubClient:
@@ -19,6 +21,44 @@ class GitHubClient:
         response = requests.get(url, params=params, headers=self.headers)
         response.raise_for_status()
         return response.json()
+
+
+def sync_releases(versions: Iterable[str]) -> None:
+    """
+    All given versions have a release. If a release is absent, persisted
+    releases are out of date and need to have their release flag removed.
+    Finally the latest version must be marked as current.
+    :param versions: an iterable yielding all retrieved GitHub tags
+    :return:
+    """
+    current_releases = NextcloudRelease.objects.all()
+    imported_releases = [
+        NextcloudRelease.objects.get_or_create(version=version)[0]
+        for version in versions
+    ]
+    if imported_releases:
+        # all imported releases have a release, existing ones don't
+        for release in imported_releases:
+            release.has_release = True
+            release.save()
+        for release in get_old_releases(current_releases, imported_releases):
+            release.has_release = False
+            release.save()
+        # set latest release
+        NextcloudRelease.objects.update(is_current=False)
+        latest = max(imported_releases, key=lambda v: Version(v.version))
+        latest.is_current = True
+        latest.save()
+
+
+NextcloudReleases = List[NextcloudRelease]
+
+
+def get_old_releases(current: NextcloudReleases,
+                     imported: NextcloudReleases) -> NextcloudReleases:
+    imported_versions = {release.version for release in imported}
+    return [release for release in current
+            if release.version not in imported_versions]
 
 
 def get_supported_releases(client: GitHubClient,